refactor(pages): rename index files to page names
Rename generic index.tsx files to descriptive kebab-case filenames to improve project structure and searchability. All corresponding imports in popup and sidepanel entrypoints have been updated.
