+++ src/entrypoints/sidepanel/pages/entries/[id]/entry-detail.tsx
