+++ src/entrypoints/sidepanel/pages/entries/edit/[id]/entry-edit.tsx
