+++ src/entrypoints/sidepanel/pages/entries/entries.tsx
