+++ src/entrypoints/sidepanel/pages/entries/new/entry-create.tsx
