+++ src/entrypoints/sidepanel/pages/home/home.tsx
