+++ src/entrypoints/sidepanel/pages/settings/settings.tsx
